Refactor list items rendering
