+++ src/webclient/main/components/registration-page/index.js
@@ -23,7 +23,7 @@ export default class RegistrationPage extends Component {
    */
   _markup() {
     return `
-            <section class="container base-form login-form">
+            <section class="container base-form login-form" data-render="registration-page">
                  <header class="header">
                     <img class="logo" alt="logo" src="./static/images/teamdev.png" width="150">
                     <i class="glyphicon glyphicon-user user-icon"></i>
@@ -37,7 +37,7 @@ export default class RegistrationPage extends Component {
    @inheritdoc.
    */
   _initInnerComponents() {
-    const formRoot = this.container.querySelector('.login-form');
+    const formRoot = this.container.querySelector('[data-render="registration-page"]');
     new RegistrationForm(formRoot, {});
   }
 }
